+++ cypress/support/step-definitions/show-edit-pod-steps.js
@@ -1,16 +1,13 @@
 import {
   showEditPodEdit, showEditPodCancelButton, showEditPodSaveButton, showEditPodDeleteButton,
-  showEditPodTitle, showEditPodPreview, showEditPodSecondaryBlock,
-  showEditPodCollapsibleInnerContent, showEditPodFooter,
+  showEditPodTitle, showEditPodSecondaryBlock,
+  showEditPodCollapsibleInnerContent, showEditPodFooter, showEditPodComponent,
 } from '../../locators/show-edit-pod';
 import { icon } from '../../locators';
 
-const SHOW_EDIT_POD_BLOCK_CONTENT = 'carbon-pod__block--';
-const SHOW_EDIT_POD_CONTENT = 'carbon-pod__content--';
-const SHOW_EDIT_POD_NO_BORDER_CONTENT = 'carbon-pod--no-border';
-const FIRST_DIV = 1;
-const SECOND_DIV = 2;
-const THIRD_DIV = 3;
+const FIRST_DIV = 0;
+const SECOND_DIV = 1;
+const THIRD_DIV = 2;
 const INNER_CONTENT_TITLE = 'title';
 const INNER_CONTENT_BODY = 'body';
 
@@ -38,20 +35,11 @@ When('Edit icon has color {string}', (color) => {
   icon().should('have.css', 'color', color);
 });
 
-Then('Show Edit Pod component has border property', () => {
-  showEditPodPreview()
-    .should('not.have.class', `${SHOW_EDIT_POD_BLOCK_CONTENT}no-border`)
-    .and('have.css', 'border', '1px solid rgb(204, 214, 219)');
-  showEditPodPreview().children()
-    .should('not.have.class', SHOW_EDIT_POD_NO_BORDER_CONTENT);
-});
-
-Then('Show Edit Pod component has no border property', () => {
-  showEditPodPreview()
-    .should('have.class', `${SHOW_EDIT_POD_BLOCK_CONTENT}no-border`)
-    .and('have.css', 'border', '0px none rgba(0, 0, 0, 0.85)');
-  showEditPodPreview().children()
-    .should('have.class', SHOW_EDIT_POD_NO_BORDER_CONTENT);
+Then('Show Edit Pod component has border {string} color', (color) => {
+  showEditPodComponent().find('div').eq(FIRST_DIV).should('have.css', 'border-bottom-color', color)
+    .and('have.css', 'border-left-color', color)
+    .and('have.css', 'border-right-color', color)
+    .and('have.css', 'border-top-color', color);
 });
 
 Then('Show Edit Pod component on a secondary block has border property', () => {
@@ -63,8 +51,7 @@ Then('Show Edit Pod component on a secondary block has no border property', () =
 });
 
 Then('Show Edit Pod component cancel button has color {string} and borderColor {string}', (color, borderColor) => {
-  showEditPodCancelButton()
-    .should('be.visible')
+  showEditPodCancelButton().should('be.visible')
     .and('have.css', 'background', 'rgba(0, 0, 0, 0) none repeat scroll 0% 0% / auto padding-box border-box')
     .and('have.css', 'border-color', borderColor)
     .and('have.css', 'color', color);
@@ -75,32 +62,27 @@ Then('Show Edit Pod component hasn\'t a cancel button', () => {
 });
 
 Then('Show Edit Pod classic component has saving property', () => {
-  showEditPodSaveButton().should('be.disabled');
-  showEditPodSaveButton().should('have.attr', 'disabled');
-  showEditPodSaveButton()
-    .should('have.css', 'background', 'rgb(230, 235, 237) none repeat scroll 0% 0% / auto padding-box border-box')
+  showEditPodSaveButton().should('be.disabled')
+    .and('have.css', 'background', 'rgb(230, 235, 237) none repeat scroll 0% 0% / auto padding-box border-box')
     .and('have.css', 'border', '1px solid rgba(0, 0, 0, 0)')
-    .and('have.css', 'color', 'rgba(0, 0, 0, 0.2)');
+    .and('have.css', 'color', 'rgba(0, 0, 0, 0.2)')
+    .and('have.attr', 'disabled');
 });
 
 Then('Show Edit Pod component has saving property', () => {
-  showEditPodSaveButton().should('be.disabled');
-  showEditPodSaveButton().should('have.attr', 'disabled');
-  showEditPodSaveButton()
-    .should('have.css', 'background', 'rgb(229, 234, 236) none repeat scroll 0% 0% / auto padding-box border-box')
+  showEditPodSaveButton().should('be.disabled')
+    .and('have.css', 'background', 'rgb(229, 234, 236) none repeat scroll 0% 0% / auto padding-box border-box')
     .and('have.css', 'border', '2px solid rgba(0, 0, 0, 0)')
-    .and('have.css', 'color', 'rgba(0, 0, 0, 0.3)');
+    .and('have.css', 'color', 'rgba(0, 0, 0, 0.3)')
+    .and('have.attr', 'disabled');
 });
 
 Then('Show Edit Pod component has no saving property', () => {
   showEditPodSaveButton().should('not.have.attr', 'disabled');
 });
 
-Then('Show Edit Pod as value is set to {string}', (property) => {
-  showEditPodPreview()
-    .should('have.class', `${SHOW_EDIT_POD_BLOCK_CONTENT}${property}`);
-  showEditPodPreview().children()
-    .should('have.class', `${SHOW_EDIT_POD_CONTENT}${property}`);
+Then('Show Edit Pod background-color is set to {string}', (color) => {
+  showEditPodComponent().find('div').eq(FIRST_DIV).should('have.css', 'background-color', color);
 });
 
 Then('Show Edit Pod buttons are aligned to {string}', (position) => {
@@ -120,18 +102,11 @@ When('I click cancel button', () => {
 });
 
 Then('Show Edit Pod component has proper content inside itself', () => {
-  showEditPodTitle()
-    .should('have.text', 'Person');
-  showEditPodCollapsibleInnerContent(FIRST_DIV, INNER_CONTENT_TITLE)
-    .should('have.text', 'First Name');
-  showEditPodCollapsibleInnerContent(FIRST_DIV, INNER_CONTENT_BODY)
-    .should('have.text', 'Alan');
-  showEditPodCollapsibleInnerContent(SECOND_DIV, INNER_CONTENT_TITLE)
-    .should('have.text', 'Last Name');
-  showEditPodCollapsibleInnerContent(SECOND_DIV, INNER_CONTENT_BODY)
-    .should('have.text', 'Smith');
-  showEditPodCollapsibleInnerContent(THIRD_DIV, INNER_CONTENT_TITLE)
-    .should('have.text', 'Telephone');
-  showEditPodCollapsibleInnerContent(THIRD_DIV, INNER_CONTENT_BODY)
-    .should('have.text', '000 000 0000');
+  showEditPodTitle().should('have.text', 'Person');
+  showEditPodCollapsibleInnerContent(INNER_CONTENT_TITLE, FIRST_DIV).should('have.text', 'First Name');
+  showEditPodCollapsibleInnerContent(INNER_CONTENT_BODY, FIRST_DIV).should('have.text', 'Alan');
+  showEditPodCollapsibleInnerContent(INNER_CONTENT_TITLE, SECOND_DIV).should('have.text', 'Last Name');
+  showEditPodCollapsibleInnerContent(INNER_CONTENT_BODY, SECOND_DIV).should('have.text', 'Smith');
+  showEditPodCollapsibleInnerContent(INNER_CONTENT_TITLE, THIRD_DIV).should('have.text', 'Telephone');
+  showEditPodCollapsibleInnerContent(INNER_CONTENT_BODY, THIRD_DIV).should('have.text', '000 000 0000');
 });