+++ src/style/themes/base/base-theme.config.js
@@ -107,6 +107,10 @@ export default (palette) => {
       switch: palette.slateTint(89.5)
     },
 
+    draggableItem: {
+      border: palette.slateTint(90)
+    },
+
     checkable: {
       checked: 'rgba(0,0,0,0.90)'
     },